requirements.txt
pytest==7.4.0
xmltodict==0.13.0
pytest-xdist==3.3.1
selenium==4.10.0
webdriver-manager==4.0.0
requests==2.31.0
python-dotenv==1.0.0